Allow sending messages directly from driver (FreeRTOS) tasks
Introduce and document new functions to send messages from a driver task,
typically a callback on a platform with FreeRTOS.

While sending messages from such a task may work with SMP builds, it was not
necessarily safe and it didn't work with non-SMP builds.

Introduce the required synchronization primitives that try and acquire locks.

Also fix ESP32 network driver to use these new functions as it was broken with
non-SMP builds (e.g. ESP32C6).

Also fix emscripten/cypress test failure that may be related to
https://github.com/cypress-io/cypress/issues/20802

Signed-off-by: Paul Guyot <pguyot@kallisys.net>